Split up span summary page components
